# docs/openapi.yaml
openapi: 3.1.0
info:
  title: HelpDesk API (AS-IS)
  version: 0.1.0
  description: OpenAPI baseline reflecting currently implemented routes.
servers:
  - url: https://example.com
    description: Replace with deployment host
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
    ErrorAsIs:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [REQUESTER, AGENT, ADMIN]
        organizationId:
          type: string
          format: uuid
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        ticketCount:
          type: integer
          description: Total tickets created by this user
        activeTicketCount:
          type: integer
          description: Active tickets assigned to this user
      additionalProperties: true
    TeamSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        organizationId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        memberCount:
          type: integer
          description: Number of team members
        activeTicketCount:
          type: integer
          description: Active tickets assigned to this team
      additionalProperties: true
    Ticket:
      type: object
      properties:
        id:
          type: string
          format: uuid
        number:
          type: integer
        title:
          type: string
        descriptionMd:
          type: string
        status:
          type: string
          enum: [NOWE, W_TOKU, OCZEKUJE_NA_UZYTKOWNIKA, WSTRZYMANE, ROZWIAZANE, ZAMKNIETE, PONOWNIE_OTWARTE]
        priority:
          type: string
          enum: [NISKI, SREDNI, WYSOKI, KRYTYCZNY]
        category:
          type: string
          nullable: true
        requesterId:
          type: string
          format: uuid
        assigneeUserId:
          type: string
          format: uuid
          nullable: true
        assigneeTeamId:
          type: string
          format: uuid
          nullable: true
        organizationId:
          type: string
          format: uuid
        firstResponseAt:
          type: string
          format: date-time
          nullable: true
        firstResponseDue:
          type: string
          format: date-time
          nullable: true
        resolveDue:
          type: string
          format: date-time
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        closedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TicketExpanded:
      allOf:
        - $ref: '#/components/schemas/Ticket'
        - type: object
          properties:
            requester:
              $ref: '#/components/schemas/UserSummary'
            assigneeUser:
              $ref: '#/components/schemas/UserSummary'
            assigneeTeam:
              $ref: '#/components/schemas/TeamSummary'
    TicketCreate:
      type: object
      required: [title, descriptionMd, priority]
      properties:
        title:
          type: string
          minLength: 3
        descriptionMd:
          type: string
          minLength: 3
        priority:
          type: string
          enum: [NISKI, SREDNI, WYSOKI, KRYTYCZNY]
        category:
          type: string
          nullable: true
...
paths:
  /api/tickets:
    get:
      summary: List tickets
      description: |
        Returns tickets scoped by user role and organization.
        
        - REQUESTER role: Returns only tickets created by the user
        - AGENT/ADMIN roles: Returns tickets within the user's organization
        
        Supports filtering by status, priority, category, tags, date ranges, assignees, and text search.
        Uses cursor-based pagination for efficient navigation.
      security:
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          description: "Number of tickets per page (1-100, default: 20)"
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          description: Base64-encoded cursor for pagination
          schema:
            type: string
        - name: direction
          in: query
          description: Pagination direction
          schema:
            type: string
            enum: [next, prev]
            default: next
        - name: status
          in: query
          description: Filter by ticket status
          schema:
            type: string
            enum: [NOWE, W_TOKU, OCZEKUJE_NA_UZYTKOWNIKA, WSTRZYMANE, ROZWIAZANE, ZAMKNIETE, PONOWNIE_OTWARTE]
        - name: priority
          in: query
          description: Filter by ticket priority
          schema:
            type: string
            enum: [NISKI, SREDNI, WYSOKI, KRYTYCZNY]
        - name: q
          in: query
          description: Search in title and description (case-insensitive)
          schema:
            type: string
        - name: category
          in: query
          description: Filter by category ID
          schema:
            type: string
        - name: tags
          in: query
          description: Tag IDs (comma-separated)
          schema:
            type: string
        - name: createdAtFrom
          in: query
          description: Filter tickets created on or after this date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: createdAtTo
          in: query
          description: Filter tickets created on or before this date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: updatedAtFrom
          in: query
          description: Filter tickets updated on or after this date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: updatedAtTo
          in: query
          description: Filter tickets updated on or before this date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: resolvedAtFrom
          in: query
          description: Filter tickets resolved on or after this date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: resolvedAtTo
          in: query
          description: Filter tickets resolved on or before this date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: assigneeUserId
          in: query
          description: Filter tickets assigned to a specific user
          schema:
            type: string
            format: uuid
        - name: assigneeTeamId
          in: query
          description: Filter tickets assigned to a specific team
          schema:
            type: string
            format: uuid
        - name: sortBy
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [createdAt, updatedAt, resolvedAt, priority, status]
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Tickets list
          content:
            application/json:
              schema:
                type: object
                properties:
                  tickets:
                    type: array
                    items:
                      $ref: '#/components/schemas/TicketExpanded'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
    post:
      summary: Create ticket
      description: |
        Creates a new ticket with the provided title, description, and priority.
        
        Automatically:
        - Sets requester to the authenticated user
        - Assigns to the user's organization
        - Calculates SLA due dates based on priority and category
        - Schedules SLA breach detection jobs
        - Creates initial audit event
        
        Requires: title (min 3 chars), descriptionMd (min 3 chars), priority.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreate'
      responses:
        '200':
          description: Ticket created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket:
                    $ref: '#/components/schemas/Ticket'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorAsIs'
...